This is a simple portfolio website.
The Personal Portfolio Website project is designed to showcase an individual's skills, projects, experience, and personal information in a visually appealing and organized manner. Built using HTML and CSS, this portfolio serves as an online resume and professional presence, allowing visitors to gain insight into the individual's technical capabilities and achievements.